# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Campus Complaint Desk API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  /** @nullable */
  adminCode?: string | null;
  /** @nullable */
  staffCode?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  staff: "staff",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type ChangeRoleBodyRole =
  (typeof ChangeRoleBodyRole)[keyof typeof ChangeRoleBodyRole];

export const ChangeRoleBodyRole = {
  student: "student",
  staff: "staff",
  admin: "admin",
} as const;

export interface ChangeRoleBody {
  role: ChangeRoleBodyRole;
}

export type CreateComplaintBodyCategoryUser =
  (typeof CreateComplaintBodyCategoryUser)[keyof typeof CreateComplaintBodyCategoryUser];

export const CreateComplaintBodyCategoryUser = {
  maintenance: "maintenance",
  hygiene: "hygiene",
  food: "food",
  internet: "internet",
  security: "security",
  noise: "noise",
  other: "other",
} as const;

export interface CreateComplaintBody {
  title: string;
  description: string;
  categoryUser: CreateComplaintBodyCategoryUser;
  location: string;
  /** @nullable */
  imageUrl?: string | null;
}

/**
 * @nullable
 */
export type UpdateComplaintBodyStatus =
  | (typeof UpdateComplaintBodyStatus)[keyof typeof UpdateComplaintBodyStatus]
  | null;

export const UpdateComplaintBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  resolved: "resolved",
} as const;

export interface UpdateComplaintBody {
  /** @nullable */
  status?: UpdateComplaintBodyStatus;
  /** @nullable */
  feedback?: string | null;
  /** @nullable */
  feedbackRating?: number | null;
}

export interface AssignComplaintBody {
  staffId: number;
}

export type ComplaintStatus =
  (typeof ComplaintStatus)[keyof typeof ComplaintStatus];

export const ComplaintStatus = {
  pending: "pending",
  in_progress: "in_progress",
  resolved: "resolved",
} as const;

export interface Complaint {
  id: number;
  title: string;
  description: string;
  categoryUser: string;
  /** @nullable */
  categoryAi?: string | null;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  sentiment?: string | null;
  status: ComplaintStatus;
  /** @nullable */
  imageUrl?: string | null;
  location: string;
  /** @nullable */
  feedback?: string | null;
  /** @nullable */
  feedbackRating?: number | null;
  userId: number;
  /** @nullable */
  assignedTo?: number | null;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  assignedToName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ComplaintStatsByCategoryItem = {
  category: string;
  count: number;
};

export type ComplaintStatsByPriorityItem = {
  priority: string;
  count: number;
};

export interface ComplaintStats {
  total: number;
  pending: number;
  inProgress: number;
  resolved: number;
  byCategory: ComplaintStatsByCategoryItem[];
  byPriority: ComplaintStatsByPriorityItem[];
}

export type GetComplaintsParams = {
  status?: string;
  category?: string;
};